refactor: Enemy Entity
Decoupled the logic layer and data layer.

Use Timer instead of delayed execution in classes that do not inherit Mono (the effect is problematic and needs to be studied)

-

undone:

Refactoring of SoundSpread.

Part of the code (especially Filed, Trigger, etc.) is not elegant enough and needs to be refactored.